Renamed and moved ConventionResolver to JavaConventionResolver
